LPS-55831 moves classes
