+++ portal-impl/src/com/liferay/portlet/exportconfiguration/action/EditPublishConfigurationAction.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.layoutsadmin.action;
+package com.liferay.portlet.exportconfiguration.action;
 
 import com.liferay.portal.NoSuchGroupException;
 import com.liferay.portal.kernel.exception.PortalException;
@@ -23,28 +23,23 @@ import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.servlet.SessionErrors;
 import com.liferay.portal.kernel.servlet.SessionMessages;
+import com.liferay.portal.kernel.staging.StagingUtil;
 import com.liferay.portal.kernel.util.Constants;
 import com.liferay.portal.kernel.util.MapUtil;
 import com.liferay.portal.kernel.util.ParamUtil;
-import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.model.BackgroundTask;
 import com.liferay.portal.model.ExportImportConfiguration;
-import com.liferay.portal.model.TrashedModel;
 import com.liferay.portal.security.auth.PrincipalException;
 import com.liferay.portal.service.BackgroundTaskLocalServiceUtil;
-import com.liferay.portal.service.ExportImportConfigurationServiceUtil;
-import com.liferay.portal.service.ExportImportServiceUtil;
-import com.liferay.portal.struts.PortletAction;
+import com.liferay.portal.service.ExportImportConfigurationLocalServiceUtil;
+import com.liferay.portal.theme.ThemeDisplay;
 import com.liferay.portal.util.PortalUtil;
+import com.liferay.portal.util.WebKeys;
 import com.liferay.portlet.sites.action.ActionUtil;
-import com.liferay.portlet.trash.service.TrashEntryServiceUtil;
-import com.liferay.portlet.trash.util.TrashUtil;
 
 import java.io.Serializable;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Map;
 
 import javax.portlet.ActionRequest;
@@ -62,9 +57,12 @@ import org.apache.struts.action.ActionForward;
 import org.apache.struts.action.ActionMapping;
 
 /**
+ * @author Daniel Kocsis
+ * @author Mate Thurzo
  * @author Levente Hudák
  */
-public class EditExportConfigurationAction extends PortletAction {
+public class EditPublishConfigurationAction
+	extends EditExportConfigurationAction {
 
 	@Override
 	public void processAction(
@@ -77,28 +75,32 @@ public class EditExportConfigurationAction extends PortletAction {
 
 		String cmd = ParamUtil.getString(actionRequest, Constants.CMD);
 
+		ThemeDisplay themeDisplay = (ThemeDisplay)actionRequest.getAttribute(
+			WebKeys.THEME_DISPLAY);
+
 		try {
+			long exportImportConfigurationId = ParamUtil.getLong(
+				actionRequest, "exportImportConfigurationId");
+
 			if (cmd.equals(Constants.ADD) || cmd.equals(Constants.UPDATE)) {
-				updateExportConfiguration(actionRequest);
+				updatePublishConfiguration(actionRequest);
 			}
 			else if (cmd.equals(Constants.DELETE)) {
 				deleteExportImportConfiguration(actionRequest, false);
 			}
-			else if (cmd.equals(Constants.EXPORT)) {
-				long exportImportConfigurationId = ParamUtil.getLong(
-					actionRequest, "exportImportConfigurationId");
-
-				ExportImportServiceUtil.exportLayoutsAsFileInBackground(
-					exportImportConfigurationId);
-			}
 			else if (cmd.equals(Constants.MOVE_TO_TRASH)) {
 				deleteExportImportConfiguration(actionRequest, true);
 			}
-			else if (cmd.equals(Constants.RESTORE)) {
-				restoreTrashEntries(actionRequest);
+			else if (cmd.equals(Constants.PUBLISH_TO_LIVE)) {
+				StagingUtil.publishLayouts(
+					themeDisplay.getUserId(), exportImportConfigurationId);
+			}
+			else if (cmd.equals(Constants.PUBLISH_TO_REMOTE)) {
+				StagingUtil.copyRemoteLayouts(exportImportConfigurationId);
 			}
 			else if (cmd.equals(Constants.RELAUNCH)) {
-				relaunchExportLayoutConfiguration(actionRequest);
+				relaunchPublishLayoutConfiguration(
+					themeDisplay.getUserId(), actionRequest);
 			}
 			else if (Validator.isNull(cmd)) {
 				addSessionMessages(actionRequest);
@@ -133,7 +135,8 @@ public class EditExportConfigurationAction extends PortletAction {
 
 				SessionErrors.add(renderRequest, e.getClass());
 
-				return actionMapping.findForward("portlet.layouts_admin.error");
+				return actionMapping.findForward(
+					"portlet.export_configuration.error");
 			}
 			else {
 				throw e;
@@ -141,7 +144,8 @@ public class EditExportConfigurationAction extends PortletAction {
 		}
 
 		return actionMapping.findForward(
-			getForward(renderRequest, "portlet.layouts_admin.export_layouts"));
+			getForward(
+				renderRequest, "portlet.export_configuration.publish_layouts"));
 	}
 
 	@Override
@@ -155,11 +159,13 @@ public class EditExportConfigurationAction extends PortletAction {
 
 		PortletRequestDispatcher portletRequestDispatcher =
 			portletContext.getRequestDispatcher(
-				"/html/portlet/layouts_admin/export_layouts_processes.jsp");
+				"/html/portlet/export_configuration/" +
+					"publish_layouts_processes.jsp");
 
 		portletRequestDispatcher.include(resourceRequest, resourceResponse);
 	}
 
+	@Override
 	protected void addSessionMessages(ActionRequest actionRequest)
 		throws Exception {
 
@@ -180,64 +186,28 @@ public class EditExportConfigurationAction extends PortletAction {
 
 		long groupId = ParamUtil.getLong(actionRequest, "groupId");
 
+		int exportImportConfigurationType =
+			ExportImportConfigurationConstants.TYPE_PUBLISH_LAYOUT_REMOTE;
+
+		boolean localPublishing = ParamUtil.getBoolean(
+			actionRequest, "localPublishing");
+
+		if (localPublishing) {
+			exportImportConfigurationType =
+				ExportImportConfigurationConstants.TYPE_PUBLISH_LAYOUT_LOCAL;
+		}
+
 		Map<String, Serializable> settingsMap =
 			ExportImportConfigurationSettingsMapFactory.buildSettingsMap(
-				actionRequest, groupId,
-				ExportImportConfigurationConstants.TYPE_EXPORT_LAYOUT);
+				actionRequest, groupId, exportImportConfigurationType);
 
 		SessionMessages.add(
 			actionRequest, portletId + "settingsMap", settingsMap);
 	}
 
-	protected void deleteExportImportConfiguration(
-			ActionRequest actionRequest, boolean moveToTrash)
+	protected void relaunchPublishLayoutConfiguration(
+			long userId, ActionRequest actionRequest)
 		throws PortalException {
-
-		long[] deleteExportImportConfigurationIds = null;
-
-		long exportImportConfigurationId = ParamUtil.getLong(
-			actionRequest, "exportImportConfigurationId");
-
-		if (exportImportConfigurationId > 0) {
-			deleteExportImportConfigurationIds =
-				new long[] {exportImportConfigurationId};
-		}
-		else {
-			deleteExportImportConfigurationIds = StringUtil.split(
-				ParamUtil.getString(
-					actionRequest, "deleteExportImportConfigurationIds"), 0L);
-		}
-
-		List<TrashedModel> trashedModels = new ArrayList<>();
-
-		for (long deleteExportImportConfigurationId :
-				deleteExportImportConfigurationIds) {
-
-			if (moveToTrash) {
-				ExportImportConfiguration exportImportConfiguration =
-					ExportImportConfigurationServiceUtil.
-						moveExportImportConfigurationToTrash(
-							deleteExportImportConfigurationId);
-
-				trashedModels.add(exportImportConfiguration);
-			}
-			else {
-				ExportImportConfigurationServiceUtil.
-					deleteExportImportConfiguration(
-						deleteExportImportConfigurationId);
-			}
-		}
-
-		if (moveToTrash && !trashedModels.isEmpty()) {
-			TrashUtil.addTrashSessionMessages(actionRequest, trashedModels);
-
-			hideDefaultSuccessMessage(actionRequest);
-		}
-	}
-
-	protected void relaunchExportLayoutConfiguration(
-			ActionRequest actionRequest)
-		throws Exception {
 
 		long backgroundTaskId = ParamUtil.getLong(
 			actionRequest, "backgroundTaskId");
@@ -248,42 +218,62 @@ public class EditExportConfigurationAction extends PortletAction {
 		Map<String, Serializable> taskContextMap =
 			backgroundTask.getTaskContextMap();
 
-		long exportImportConfigurationId = MapUtil.getLong(
-			taskContextMap, "exportImportConfigurationId");
+		ExportImportConfiguration exportImportConfiguration =
+			ExportImportConfigurationLocalServiceUtil.
+				getExportImportConfiguration(
+					MapUtil.getLong(
+						taskContextMap, "exportImportConfigurationId"));
 
-		ExportImportServiceUtil.exportLayoutsAsFileInBackground(
-			exportImportConfigurationId);
-	}
+		if (exportImportConfiguration.getType() ==
+				ExportImportConfigurationConstants.TYPE_PUBLISH_LAYOUT_LOCAL) {
 
-	protected void restoreTrashEntries(ActionRequest actionRequest)
-		throws Exception {
+			StagingUtil.publishLayouts(userId, exportImportConfiguration);
+		}
+		else if (exportImportConfiguration.getType() ==
+					ExportImportConfigurationConstants.
+						TYPE_PUBLISH_LAYOUT_REMOTE) {
 
-		long[] restoreTrashEntryIds = StringUtil.split(
-			ParamUtil.getString(actionRequest, "restoreTrashEntryIds"), 0L);
-
-		for (long restoreTrashEntryId : restoreTrashEntryIds) {
-			TrashEntryServiceUtil.restoreEntry(restoreTrashEntryId);
+			StagingUtil.copyRemoteLayouts(exportImportConfiguration);
 		}
 	}
 
-	protected ExportImportConfiguration updateExportConfiguration(
+	protected ExportImportConfiguration updatePublishConfiguration(
 			ActionRequest actionRequest)
 		throws Exception {
 
 		long exportImportConfigurationId = ParamUtil.getLong(
 			actionRequest, "exportImportConfigurationId");
 
+		boolean localPublishing = ParamUtil.getBoolean(
+			actionRequest, "localPublishing");
+
 		if (exportImportConfigurationId > 0) {
-			return ExportImportConfigurationHelper.
-				updateExportLayoutExportImportConfiguration(actionRequest);
+			if (localPublishing) {
+				return ExportImportConfigurationHelper.
+					updatePublishLayoutLocalExportImportConfiguration(
+						actionRequest);
+			}
+			else {
+				return ExportImportConfigurationHelper.
+					updatePublishLayoutRemoteExportImportConfiguration(
+						actionRequest);
+			}
 		}
 		else {
-			return ExportImportConfigurationHelper.
-				addExportLayoutExportImportConfiguration(actionRequest);
+			if (localPublishing) {
+				return ExportImportConfigurationHelper.
+					addPublishLayoutLocalExportImportConfiguration(
+						actionRequest);
+			}
+			else {
+				return ExportImportConfigurationHelper.
+					addPublishLayoutRemoteExportImportConfiguration(
+						actionRequest);
+			}
 		}
 	}
 
 	private static final Log _log = LogFactoryUtil.getLog(
-		EditExportConfigurationAction.class);
+		EditPublishConfigurationAction.class);
 
 }
