+++ portal-impl/src/com/liferay/portlet/exportconfiguration/messaging/DraftExportImportConfigurationMessageListener.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.layoutsadmin.messaging;
+package com.liferay.portlet.exportconfiguration.messaging;
 
 import com.liferay.portal.kernel.dao.orm.DynamicQuery;
 import com.liferay.portal.kernel.dao.orm.Order;
