+++ portal-impl/src/com/liferay/portlet/exportconfiguration/search/ExportImportConfigurationDisplayTerms.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.layoutsadmin.search;
+package com.liferay.portlet.exportconfiguration.search;
 
 import com.liferay.portal.kernel.dao.search.DisplayTerms;
 import com.liferay.portal.kernel.util.ParamUtil;
