+++ portal-impl/src/com/liferay/portlet/exportconfiguration/trash/ExportImportConfigurationTrashHandler.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.layoutsadmin.trash;
+package com.liferay.portlet.exportconfiguration.trash;
 
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.trash.BaseTrashHandler;
