+++ portal-impl/src/com/liferay/portlet/exportconfiguration/util/ExportImportConfigurationIndexer.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.layoutsadmin.util;
+package com.liferay.portlet.exportconfiguration.util;
 
 import com.liferay.portal.kernel.dao.orm.ActionableDynamicQuery;
 import com.liferay.portal.kernel.exception.PortalException;
